Change lowest number calculation to use integers
Changed the data type from float to int for lowest number comparison.

+++ find_lowest_number.py
@@ -24,13 +24,13 @@ with open(input_filename, 'r') as input_file:
     for line in input_file:
         if number_found == False:
             try:
-                lowest_number = float(line)
+                lowest_number = int(line)
                 number_found = True
             except ValueError:
                 break
         else:
-            if float(line) < lowest_number:
-                lowest_number = float(line)
+            if int(line) < lowest_number:
+                lowest_number = int(line)
 
 with open(output_filename, 'w') as output_file:
     if number_found:
